requirements.txt
customtkinter
mutagen
Pillow
piexif
pypdf
python-docx
openpyxl
pyinstaller
zstandard
